Match itfPanel callback parameter widths

+++ src/itfPanel.c
@@ -159,8 +159,8 @@ u64 FUN_003ad400(u16 param_1,u16 param_2,u64 param_3,u64 param_4);
 extern u64 FUN_003ad400_typed(u16 param_1,u16 param_2,u64 param_3,int param_4);
 u64 FUN_003ad640(u64 param_1,u64 param_2);
 u64 FUN_003ad740(u64 param_1,int param_2);
-u64 FUN_003ad860(u64 param_1,u64 param_2);
-u64 FUN_003ad930(u64 param_1,u64 param_2);
+u64 FUN_003ad860(u64 param_1,int param_2);
+u64 FUN_003ad930(u64 param_1,int param_2);
 u64 FUN_003ada00(u64 param_1,u64 param_2);
 u64 FUN_003adb00(u64 param_1,int param_2);
 u64 FUN_003adb80(u64 param_1,int param_2);
@@ -4019,7 +4019,7 @@ u64 FUN_003ad740(u64 param_1,int param_2)
 // FUN_003AD860 NONMATCHING
 
 
-u64 FUN_003ad860(u64 param_1,u64 param_2)
+u64 FUN_003ad860(u64 param_1,int param_2)
 
 
 
@@ -4081,7 +4081,7 @@ u64 FUN_003ad860(u64 param_1,u64 param_2)
 // FUN_003AD930 NONMATCHING
 
 
-u64 FUN_003ad930(u64 param_1,u64 param_2)
+u64 FUN_003ad930(u64 param_1,int param_2)
